requirements.txt
asgiref==3.2.7
azure-common==1.1.27
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==2.2.20
django-cors-headers==3.7.0
django-heroku==0.3.1
django-import-export==2.5.0
django-storages==1.11.1
django-suit==0.2.28
django-versatileimagefield==2.1
djangorestframework==3.11.0
et-xmlfile==1.1.0
gunicorn==20.0.4
idna==3.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
psycopg2-binary==2.8.6
pycparser==2.20
python-dateutil==2.8.1
python-magic==0.4.24
pytz==2020.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
sqlparse==0.3.1
tablib==3.0.0
threadpoolctl==2.1.0
urllib3==1.26.6
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
